+++ plugin/galerie/param/routes.php
@@ -12,3 +12,10 @@ defined('ROOT') OR exit('No direct script access allowed');
 $router = router::getInstance();
 
 $router->map('GET', '/galerie[/?]', 'GalerieController#home', 'galerie-home');
+
+$router->map('GET', '/admin/galerie[/?]', 'GalerieAdminController#home', 'admin-galerie-home');
+$router->map('GET', '/admin/galerie/edit', 'GalerieAdminController#edit', 'admin-galerie-edit');
+$router->map('GET', '/admin/galerie/edit/[a:id]', 'GalerieAdminController#editId', 'admin-galerie-edit-id');
+$router->map('GET', '/admin/galerie/delete/[a:id]/[a:token]', 'GalerieAdminController#delete', 'admin-galerie-delete');
+$router->map('POST', '/admin/galerie/save', 'GalerieAdminController#save', 'admin-galerie-save');
+$router->map('POST', '/admin/galerie/saveConf', 'GalerieAdminController#saveConf', 'admin-galerie-save-config');